Added basename to dither filenames and detections, to differentiate between different virus00000?? directories if multiple are reduced in the same folder.
Removed all dependence on the IFU_DICT for the case when usemapping=False. If usemapping=True,
the CAM dict is now generated from the IFU_DICT.

Dither files are only created once now. If one already exists in the file system, there is
now an option to remake it.

The choise of bias directory is now a command line option.